Add ability to add extra ENV vars to containers
The http-add-on chart did not have the ability yet to add custom extra
env variables to the deployments. This implements it according to the
standard set in the PR implementing the extra container ags,
using an object instead of a list of objects. (See for this PR:
https://github.com/kedacore/charts/pull/245)

A previous PR was already made some time ago but it did not adhere to
the requested convention: https://github.com/kedacore/charts/pull/762

This PR tries to solve that issue.

Signed-off-by: Robin de Rooij <robin.de.rooij@rosemanlabs.com>